+++ libs/ui-article-delivery/tailwind.config.js
@@ -4,8 +4,8 @@ const { join } = require('path')
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   mode: 'jit',
-  darkMode: ['dark'],
   prefix: 'article-',
+  darkMode: ['class', '[class="dark"]'],
   content: [
     join(__dirname, '{src,pages,components,app}/**/*!(*.stories|*.spec).{ts,tsx,html}'),
     ...createGlobPatternsForDependencies(__dirname),