+++ libs/ui-card/src/components/card-coupon/card-coupon-image/card-coupon-image.tsx
@@ -1,0 +1,17 @@
+type CardCouponImageProps = {
+  logo: string
+  alt: string
+}
+
+export function CardCouponImage({ logo, alt }: CardCouponImageProps) {
+  return (
+    <img
+      className="card-mr-xxxs card-rounded-md dark:card-grayscale"
+      src={logo}
+      alt={alt}
+      loading="lazy"
+      height={58}
+      width={58}
+    />
+  )
+}